+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased]
 
+### Added
+
+- Add `hidden-features` setting to hide features from the documentation
+
 ## [1.2.0] - 2025-12-29
 
 ### Added